Extract wave campaign router models
